Update `switch` logic in Hostsfile
Make sure to close both file. The old file was never explicitly closed,
while I’m not sure if this would be a “problem”, we might as well just
close it.

Hostsfile#switchero has been renamed to just `#switch` as it seems more
fitting.

Hostsfile#switch now also applies 644 permissions for the new file,
which I guess is reasonable.